remove path validation
