+++ cmd/oras/file.go
@@ -22,16 +22,13 @@ import (
 
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 	"oras.land/oras-go/v2/content/file"
-	fileref "oras.land/oras/cmd/oras/internal/file"
+	"oras.land/oras/cmd/oras/internal/fileref"
 )
 
 func loadFiles(ctx context.Context, store *file.Store, annotations map[string]map[string]string, fileRefs []string, verbose bool) ([]ocispec.Descriptor, error) {
 	var files []ocispec.Descriptor
 	for _, fileRef := range fileRefs {
-		filename, mediaType, err := fileref.Parse(fileRef, "")
-		if err != nil {
-			return nil, err
-		}
+		filename, mediaType := fileref.Parse(fileRef, "")
 
 		// get shortest absolute path as unique name
 		name := filepath.Clean(filename)
